backend/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
sqlparse==0.5.3
tzdata==2025.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==0.21.1
